requirements.txt
crewai
fastapi
uvicorn
celery
redis
python-dotenv
langchain-community
pypdf
